BugBountyToolkit/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
colorama>=0.4.4
PySocks>=1.7.1
stem>=1.8.0
